refactor useFileHandling moved to hooks
